requirements.txt
Flask
pymongo
flasgger
python-dotenv
PyJWT
logging